Imported missing cities Several bug fixes

+++ database/migrations/2021_06_07_155611_import_barnaul_districts.php
@@ -27,6 +27,10 @@ class ImportBarnaulDistricts extends Migration
           ->where('name', $city['name'])
           ->first();
 
+        if (!$c) {
+          return;
+        }
+
         $c->districts()
           ->whereNotIn('name', $districts)
           ->forceDelete();
